Initial commit
Still a lot of debug code.

The OIG data scrape is in the oig.js file.  It does and doesn't
currently work.  It does a successful request, but results are not being
returned that should be.

Initial work on the DUNS api wsdl service is still inline on the
index.js, but not currently functional.